+++ gonk-player/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gonk-player"
+name = "gonk-old-player"
 version = "0.1.0"
 edition = "2021"
 authors = ["Bay"]
