+++ gonk/Cargo.toml
@@ -15,5 +15,5 @@ tui = "0.18.0"
 crossterm = "0.23.2"
 unicode-width = "0.1.9"
 rayon = "1.5.3"
-gonk-player = {version = "0.1.0", path = "../gonk-player"}
+gonk-player = {version = "0.1.0", path = "../player"}
 gonk-database = {version = "0.1.0", path = "../gonk-database"}
