+++ gonk/src/playlist.rs
@@ -119,7 +119,7 @@ pub fn on_enter(playlist: &mut Playlist, player: &mut Player) {
                     .map(|song| Song::from(&song.into_bytes(), 0))
                     .collect();
 
-                match player.add_songs(&songs) {
+                match player.add(&songs) {
                     Ok(_) => (),
                     Err(e) => log!("{}", e),
                 }
@@ -129,7 +129,7 @@ pub fn on_enter(playlist: &mut Playlist, player: &mut Player) {
         Mode::Song => {
             if let Some(selected) = playlist.playlists.selected() {
                 if let Some(song) = selected.songs.selected() {
-                    match player.add_songs(&[Song::from(&song.into_bytes(), 0)]) {
+                    match player.add(&[Song::from(&song.into_bytes(), 0)]) {
                         Ok(_) => (),
                         Err(e) => log!("{}", e),
                     }
