+++ player/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "player"
+name = "gonk-player"
 version = "0.1.0"
 edition = "2021"
